requirements.txt
appdirs==1.4.3
click==6.7
future==0.16.0
Jinja2==2.9.6
kraken==0.9.0
lxml==3.7.3
MarkupSafe==1.0
numpy==1.12.1
olefile==0.44
packaging==16.8
Pillow==4.1.0
protobuf==3.2.0
pyparsing==2.2.0
pytesseract==0.1.6
python-bidi==0.4.0
pyttsx==1.1
regex==2017.4.5
requests==2.13.0
scipy==0.19.0
six==1.10.0
